+++ includes/backend/class-admin-post-lists/class-admin-post-list-session.php
@@ -170,7 +170,7 @@ class Admin_Post_List_Session extends \WordPress_Helper\Admin_Post_List {
         switch ( $orderby ) {
             case 'event-date':
                 $query->set( 'orderby', 'meta_value' );
-                $query->set( 'meta_key', 'programmpunkt-datum' );
+                $query->set( 'meta_key', 'CONGRESSOMAT_SESSION_DATE_SORTKEY' );
                 break;
 
             case 'update':